registro de un nuevo topico

+++ src/main/resources/db/migration/V2__create-table-cursos.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Cursos (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id bigint AUTO_INCREMENT PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     categoria VARCHAR(100) NOT NULL
 );
